+++ src/demo/index.tsx
@@ -4,7 +4,7 @@ import { CssBaseline, ThemeProvider, createTheme } from "@material-ui/core";
 import React, { useCallback, useState } from "react";
 import ReactDom from "react-dom";
 import * as Parts from "../parts";
-import TextControlRender from "../parts/controls/TextControlRender";
+import TextControlRender from "../parts/controls/TextControlWidget";
 import * as FontNS from "../parts/font";
 import { session as dataSession } from "./data";
 import { PWD, provider } from "./interviews";