+++ src/parts/controls/ControlWidgetTypes.ts
@@ -1,6 +1,6 @@
 import { AttributeData, Control } from "@decisively-io/interview-sdk";
 
-export interface ControlRenderProps<C extends Control> {
+export interface ControlWidgetProps<C extends Control> {
   control: C;
   chOnScreenData?: (data: AttributeData) => void;
 }