+++ src/parts/controls/CurrencyControlWidget.tsx
@@ -1,4 +1,5 @@
-import { AttributeData, TextControl } from "@decisively-io/interview-sdk";
+import { AttributeData, CurrencyControl } from "@decisively-io/interview-sdk";
+import InputAdornment from "@material-ui/core/InputAdornment";
 import TextField, { TextFieldProps } from "@material-ui/core/TextField";
 import React from "react";
 import { Controller, useFormContext } from "react-hook-form";
@@ -6,50 +7,43 @@ import styled from "styled-components";
 import { deriveLabel } from "../../util/controls";
 import { InterviewContext } from "../index";
 import { DISPLAY_NAME_PREFIX } from "./ControlConstants";
-import { ControlRenderProps } from "./ControlRenderTypes";
+import { ControlWidgetProps } from "./ControlWidgetTypes";
 import FormControl from "./FormControl";
 
-export interface TextControlRenderProps extends ControlRenderProps<TextControl> {
-  textFieldProps?: TextFieldProps;
+export interface CurrencyControlWidgetProps extends ControlWidgetProps<CurrencyControl> {
+  textFieldProps?: Omit<TextFieldProps, "value">;
   className?: string;
 }
 
-type IParam = TextFieldProps;
+type IArg = { value: CurrencyControl["value"] } & NonNullable<CurrencyControlWidgetProps["textFieldProps"]>;
 
 const StyledTextField = styled(TextField)`
   flex: 1;
 `;
+const withFallback = (arg: IArg) => (arg.value === null || arg.value === undefined ? <StyledTextField {...arg} value="" /> : <StyledTextField {...arg} />);
 
-const withFallback = (arg: IParam) => (typeof arg.value === "string" ? <StyledTextField {...arg} /> : <StyledTextField {...arg} value="" />);
-
-const TextControlRender = Object.assign(
-  React.memo((props: TextControlRenderProps) => {
+const CurrencyControlWidget = Object.assign(
+  React.memo((props: CurrencyControlWidgetProps) => {
     const { control, textFieldProps, chOnScreenData, className } = props;
     const { control: formControl } = useFormContext();
-    const { attribute, multi, variation } = control;
+    const { attribute, symbol } = control;
+
+    const InputProps = React.useMemo(
+      () => ({
+        startAdornment: <InputAdornment position="start">{symbol || "$"}</InputAdornment>,
+      }),
+      [symbol],
+    );
+
     const interview = React.useContext(InterviewContext);
     const explanation = interview?.getExplanation(attribute);
-
-    const maybeWithMulti: Pick<IParam, "multiline" | "maxRows" | "minRows"> =
-      multi === undefined
-        ? {}
-        : {
-            multiline: true,
-            ...multi,
-          };
-    const maybeWithType =
-      variation === undefined
-        ? {}
-        : {
-            type: variation.type,
-          };
 
     return (
       <Controller
         control={formControl}
         name={attribute}
-        render={({ field: { onChange, value }, fieldState: { error } }) => {
-          const typedValue = value as TextControl["value"];
+        render={({ field: { value, onChange }, fieldState: { error } }) => {
+          const typedValue = value as CurrencyControl["value"];
 
           const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
             if (chOnScreenData) {
@@ -71,9 +65,8 @@ const TextControlRender = Object.assign(
                     variant: "outlined",
                     error: error !== undefined,
                     helperText: error?.message || " ",
+                    InputProps,
                     disabled: control.disabled,
-                    ...maybeWithType,
-                    ...maybeWithMulti,
                     ...textFieldProps,
                   })}
                 </>
@@ -85,14 +78,14 @@ const TextControlRender = Object.assign(
     );
   }),
   {
-    displayName: `${DISPLAY_NAME_PREFIX}/Text`,
-    /*** @deprecated use `TextInput` directly */
-    _: null as any as React.ComponentType<TextControlRenderProps>,
+    displayName: `${DISPLAY_NAME_PREFIX}/Currency`,
+    /*** @deprecated use `Currency` directly instead */
+    _: null as any as React.ComponentType<CurrencyControlWidgetProps>,
   },
 );
-TextControlRender._ = TextControlRender;
+CurrencyControlWidget._ = CurrencyControlWidget;
 
-/*** @deprecated use `TextInput` directly */
-export const _ = TextControlRender;
+/*** @deprecated use `Currency` directly instead */
+export const _ = CurrencyControlWidget;
 
-export default TextControlRender;
+export default CurrencyControlWidget;