+++ src/parts/controls/TextControlWidget.tsx
@@ -1,5 +1,4 @@
-import { AttributeData, CurrencyControl } from "@decisively-io/interview-sdk";
-import InputAdornment from "@material-ui/core/InputAdornment";
+import { AttributeData, TextControl } from "@decisively-io/interview-sdk";
 import TextField, { TextFieldProps } from "@material-ui/core/TextField";
 import React from "react";
 import { Controller, useFormContext } from "react-hook-form";
@@ -7,43 +6,50 @@ import styled from "styled-components";
 import { deriveLabel } from "../../util/controls";
 import { InterviewContext } from "../index";
 import { DISPLAY_NAME_PREFIX } from "./ControlConstants";
-import { ControlRenderProps } from "./ControlRenderTypes";
+import { ControlWidgetProps } from "./ControlWidgetTypes";
 import FormControl from "./FormControl";
 
-export interface CurrencyControlRenderProps extends ControlRenderProps<CurrencyControl> {
-  textFieldProps?: Omit<TextFieldProps, "value">;
+export interface TextControlWidgetProps extends ControlWidgetProps<TextControl> {
+  textFieldProps?: TextFieldProps;
   className?: string;
 }
 
-type IArg = { value: CurrencyControl["value"] } & NonNullable<CurrencyControlRenderProps["textFieldProps"]>;
+type IParam = TextFieldProps;
 
 const StyledTextField = styled(TextField)`
   flex: 1;
 `;
-const withFallback = (arg: IArg) => (arg.value === null || arg.value === undefined ? <StyledTextField {...arg} value="" /> : <StyledTextField {...arg} />);
 
-const CurrencyControlRender = Object.assign(
-  React.memo((props: CurrencyControlRenderProps) => {
+const withFallback = (arg: IParam) => (typeof arg.value === "string" ? <StyledTextField {...arg} /> : <StyledTextField {...arg} value="" />);
+
+const TextControlWidget = Object.assign(
+  React.memo((props: TextControlWidgetProps) => {
     const { control, textFieldProps, chOnScreenData, className } = props;
     const { control: formControl } = useFormContext();
-    const { attribute, symbol } = control;
-
-    const InputProps = React.useMemo(
-      () => ({
-        startAdornment: <InputAdornment position="start">{symbol || "$"}</InputAdornment>,
-      }),
-      [symbol],
-    );
-
+    const { attribute, multi, variation } = control;
     const interview = React.useContext(InterviewContext);
     const explanation = interview?.getExplanation(attribute);
+
+    const maybeWithMulti: Pick<IParam, "multiline" | "maxRows" | "minRows"> =
+      multi === undefined
+        ? {}
+        : {
+            multiline: true,
+            ...multi,
+          };
+    const maybeWithType =
+      variation === undefined
+        ? {}
+        : {
+            type: variation.type,
+          };
 
     return (
       <Controller
         control={formControl}
         name={attribute}
-        render={({ field: { value, onChange }, fieldState: { error } }) => {
-          const typedValue = value as CurrencyControl["value"];
+        render={({ field: { onChange, value }, fieldState: { error } }) => {
+          const typedValue = value as TextControl["value"];
 
           const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
             if (chOnScreenData) {
@@ -65,8 +71,9 @@ const CurrencyControlRender = Object.assign(
                     variant: "outlined",
                     error: error !== undefined,
                     helperText: error?.message || " ",
-                    InputProps,
                     disabled: control.disabled,
+                    ...maybeWithType,
+                    ...maybeWithMulti,
                     ...textFieldProps,
                   })}
                 </>
@@ -78,14 +85,14 @@ const CurrencyControlRender = Object.assign(
     );
   }),
   {
-    displayName: `${DISPLAY_NAME_PREFIX}/Currency`,
-    /*** @deprecated use `Currency` directly instead */
-    _: null as any as React.ComponentType<CurrencyControlRenderProps>,
+    displayName: `${DISPLAY_NAME_PREFIX}/Text`,
+    /*** @deprecated use `TextInput` directly */
+    _: null as any as React.ComponentType<TextControlWidgetProps>,
   },
 );
-CurrencyControlRender._ = CurrencyControlRender;
+TextControlWidget._ = TextControlWidget;
 
-/*** @deprecated use `Currency` directly instead */
-export const _ = CurrencyControlRender;
+/*** @deprecated use `TextInput` directly */
+export const _ = TextControlWidget;
 
-export default CurrencyControlRender;
+export default TextControlWidget;