+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.2.0
 
-- rename all control components to `XInput` instead of just `X` to avoid colliding with inbuilt names such as `Date`
+- rename all control components to `XControlWidget` instead of just `X` to avoid colliding with inbuilt names such as `Date`
 - use interview-sdk's re-export of interview types
+- add support for conditional containers